requirements/requirements.txt
opendatasets 
pandas
numpy
pillow
scikit-learn
matplotlib
torchmetrics
pytorch_lightning
chardet
torch
torchvision